source/requirements.txt
pillow
pymsgbox
pystray
